refactor: use interview dialog system for confirm
