+++ lib/interviewer/containers/Interfaces/FinishSession.js
@@ -1,23 +1,61 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 import Button from '~/lib/ui/components/Button';
-import FinishInterviewModal from '~/app/(interview)/interview/_components/FinishInterviewModal';
+import { actionCreators as dialogActions } from '../../ducks/modules/dialogs';
+import { usePathname, useRouter } from 'next/navigation';
+import { api } from '~/trpc/client';
+import { clientRevalidateTag } from '~/utils/clientRevalidate';
 
 const FinishSession = () => {
   const dispatch = useDispatch();
-  const [openFinishInterviewModal, setOpenFinishInterviewModal] =
-    useState(false);
+  const router = useRouter();
+  const pathname = usePathname();
+
+  const interviewId = pathname.split('/').pop();
+
+  const { mutateAsync: finishInterview } = api.interview.finish.useMutation({
+    onError(error) {
+      throw new Error(error.message);
+    },
+    async onSuccess() {
+      await clientRevalidateTag('interview.get.byId');
+
+      router.push('/interview/finished');
+    },
+  });
+
+  const handleConfirmFinishInterview = async () => {
+    if (!interviewId) {
+      throw new Error('No interview id found');
+    }
+    await finishInterview({ id: interviewId });
+  };
+
+  const openDialog = useCallback(
+    (dialog) => dispatch(dialogActions.openDialog(dialog)),
+    [dispatch],
+  );
 
   useEffect(() => {
     dispatch({ type: 'PLAY_SOUND', sound: 'finishSession' });
   }, [dispatch]);
 
+  const finishInterviewConfirmation = () => {
+    openDialog({
+      type: 'Confirm',
+      title: 'Are you sure you want to finish the interview?',
+      message: (
+        <p>Your responses cannot be changed after you finish the interview.</p>
+      ),
+      canCancel: true,
+      onConfirm: async () => {
+        await handleConfirmFinishInterview();
+      },
+    });
+  };
+
   return (
     <div className="interface finish-session-interface">
-      <FinishInterviewModal
-        open={openFinishInterviewModal}
-        setOpen={setOpenFinishInterviewModal}
-      />
       <div className="finish-session-interface__frame">
         <h1 className="finish-session-interface__title type--title-1">
           Finish Interview
@@ -30,9 +68,7 @@ const FinishSession = () => {
         </div>
 
         <div className="finish-session-interface__section finish-session-interface__section--buttons">
-          <Button onClick={() => setOpenFinishInterviewModal(true)}>
-            Finish
-          </Button>
+          <Button onClick={finishInterviewConfirmation}>Finish</Button>
         </div>
       </div>
     </div>
